+++ list/listlookup/types/response_test.go
@@ -1,0 +1,83 @@
+package types_test
+
+import (
+	"testing"
+
+	"github.com/michimani/gotwi/list/listlookup/types"
+	"github.com/michimani/gotwi/resources"
+	"github.com/stretchr/testify/assert"
+)
+
+func Test_ListLookupID_HasPartialError(t *testing.T) {
+	var errorTitle string = "test partical error"
+	cases := []struct {
+		name   string
+		res    *types.GetOutput
+		expect bool
+	}{
+		{
+			name: "has partical error",
+			res: &types.GetOutput{
+				Errors: []resources.PartialError{
+					{Title: &errorTitle},
+				}},
+			expect: true,
+		},
+		{
+			name: "has no partical error",
+			res: &types.GetOutput{
+				Errors: []resources.PartialError{}},
+			expect: false,
+		},
+		{
+			name: "partical error is nil",
+			res: &types.GetOutput{
+				Errors: []resources.PartialError{}},
+			expect: false,
+		},
+	}
+
+	for _, c := range cases {
+		t.Run(c.name, func(tt *testing.T) {
+			hpe := c.res.HasPartialError()
+			assert.Equal(tt, c.expect, hpe)
+		})
+	}
+}
+
+func Test_ListLookupOwnedLists_HasPartialError(t *testing.T) {
+	var errorTitle string = "test partical error"
+	cases := []struct {
+		name   string
+		res    *types.ListOwnedOutput
+		expect bool
+	}{
+		{
+			name: "has partical error",
+			res: &types.ListOwnedOutput{
+				Errors: []resources.PartialError{
+					{Title: &errorTitle},
+				}},
+			expect: true,
+		},
+		{
+			name: "has no partical error",
+			res: &types.ListOwnedOutput{
+				Errors: []resources.PartialError{}},
+			expect: false,
+		},
+		{
+			name: "partical error is nil",
+			res: &types.ListOwnedOutput{
+				Errors: []resources.PartialError{}},
+			expect: false,
+		},
+	}
+
+	for _, c := range cases {
+		t.Run(c.name, func(tt *testing.T) {
+			hpe := c.res.HasPartialError()
+			assert.Equal(tt, c.expect, hpe)
+		})
+	}
+}
