+++ list/listmember/types/parameter_test.go
@@ -1,0 +1,517 @@
+package types_test
+
+import (
+	"io"
+	"strings"
+	"testing"
+
+	"github.com/michimani/gotwi/fields"
+	"github.com/michimani/gotwi/list/listmember/types"
+	"github.com/stretchr/testify/assert"
+)
+
+func Test_ListMembersListMemberships_SetAccessToken(t *testing.T) {
+	cases := []struct {
+		name   string
+		token  string
+		expect string
+	}{
+		{
+			name:   "normal",
+			token:  "test-token",
+			expect: "test-token",
+		},
+		{
+			name:   "empty",
+			token:  "",
+			expect: "",
+		},
+	}
+
+	for _, c := range cases {
+		t.Run(c.name, func(tt *testing.T) {
+			p := &types.ListMembershipsInput{}
+			p.SetAccessToken(c.token)
+			assert.Equal(tt, c.expect, p.AccessToken())
+		})
+	}
+}
+
+func Test_ListMembersListMemberships_ResolveEndpoint(t *testing.T) {
+	const endpointRoot = "test/endpoint/"
+	const endpointBase = "test/endpoint/:id"
+
+	cases := []struct {
+		name   string
+		params *types.ListMembershipsInput
+		expect string
+	}{
+		{
+			name: "only required parameter",
+			params: &types.ListMembershipsInput{
+				ID: "lid",
+			},
+			expect: endpointRoot + "lid",
+		},
+		{
+			name: "with expansions",
+			params: &types.ListMembershipsInput{
+				ID:         "lid",
+				Expansions: fields.ExpansionList{"ex1", "ex2"},
+			},
+			expect: endpointRoot + "lid" + "?expansions=ex1%2Cex2",
+		},
+		{
+			name: "with list.fields",
+			params: &types.ListMembershipsInput{
+				ID:         "lid",
+				ListFields: fields.ListFieldList{"lf1", "lf2"},
+			},
+			expect: endpointRoot + "lid" + "?list.fields=lf1%2Clf2",
+		},
+		{
+			name: "with users.fields",
+			params: &types.ListMembershipsInput{
+				ID:         "lid",
+				UserFields: fields.UserFieldList{"uf1", "uf2"},
+			},
+			expect: endpointRoot + "lid" + "?user.fields=uf1%2Cuf2",
+		},
+		{
+			name: "with max_results",
+			params: &types.ListMembershipsInput{
+				ID:         "lid",
+				MaxResults: 10,
+			},
+			expect: endpointRoot + "lid" + "?max_results=10",
+		},
+		{
+			name: "with pagination_token",
+			params: &types.ListMembershipsInput{
+				ID:              "lid",
+				PaginationToken: "ptoken",
+			},
+			expect: endpointRoot + "lid" + "?pagination_token=ptoken",
+		},
+		{
+			name: "all query parameters",
+			params: &types.ListMembershipsInput{
+				Expansions:      fields.ExpansionList{"ex"},
+				ID:              "lid",
+				ListFields:      fields.ListFieldList{"lf"},
+				MaxResults:      10,
+				PaginationToken: "ptoken",
+				UserFields:      fields.UserFieldList{"uf"},
+			},
+			expect: endpointRoot + "lid" + "?expansions=ex&list.fields=lf&max_results=10&pagination_token=ptoken&user.fields=uf",
+		},
+		{
+			name: "has no required parameter",
+			params: &types.ListMembershipsInput{
+				Expansions:      fields.ExpansionList{"ex"},
+				UserFields:      fields.UserFieldList{"uf"},
+				ListFields:      fields.ListFieldList{"lf"},
+				MaxResults:      10,
+				PaginationToken: "pagination_token",
+			},
+			expect: "",
+		},
+	}
+
+	for _, c := range cases {
+		t.Run(c.name, func(tt *testing.T) {
+			ep := c.params.ResolveEndpoint(endpointBase)
+			assert.Equal(tt, c.expect, ep)
+		})
+	}
+}
+
+func Test_ListMembersListMemberships_Body(t *testing.T) {
+	cases := []struct {
+		name   string
+		params *types.ListMembershipsInput
+	}{
+		{
+			name:   "empty params",
+			params: &types.ListMembershipsInput{},
+		},
+		{
+			name:   "some params",
+			params: &types.ListMembershipsInput{ID: "sid"},
+		},
+	}
+
+	for _, c := range cases {
+		t.Run(c.name, func(tt *testing.T) {
+			r, err := c.params.Body()
+			assert.NoError(tt, err)
+			assert.Nil(tt, r)
+		})
+	}
+}
+
+func Test_ListMembersGet_SetAccessToken(t *testing.T) {
+	cases := []struct {
+		name   string
+		token  string
+		expect string
+	}{
+		{
+			name:   "normal",
+			token:  "test-token",
+			expect: "test-token",
+		},
+		{
+			name:   "empty",
+			token:  "",
+			expect: "",
+		},
+	}
+
+	for _, c := range cases {
+		t.Run(c.name, func(tt *testing.T) {
+			p := &types.ListInput{}
+			p.SetAccessToken(c.token)
+			assert.Equal(tt, c.expect, p.AccessToken())
+		})
+	}
+}
+
+func Test_ListMembersGet_ResolveEndpoint(t *testing.T) {
+	const endpointRoot = "test/endpoint/"
+	const endpointBase = "test/endpoint/:id"
+
+	cases := []struct {
+		name   string
+		params *types.ListInput
+		expect string
+	}{
+		{
+			name: "only required parameter",
+			params: &types.ListInput{
+				ID: "uid",
+			},
+			expect: endpointRoot + "uid",
+		},
+		{
+			name: "with expansions",
+			params: &types.ListInput{
+				ID:         "uid",
+				Expansions: fields.ExpansionList{"ex1", "ex2"},
+			},
+			expect: endpointRoot + "uid" + "?expansions=ex1%2Cex2",
+		},
+		{
+			name: "with list.fields",
+			params: &types.ListInput{
+				ID:         "uid",
+				ListFields: fields.ListFieldList{"lf1", "lf2"},
+			},
+			expect: endpointRoot + "uid" + "?list.fields=lf1%2Clf2",
+		},
+		{
+			name: "with users.fields",
+			params: &types.ListInput{
+				ID:         "uid",
+				UserFields: fields.UserFieldList{"uf1", "uf2"},
+			},
+			expect: endpointRoot + "uid" + "?user.fields=uf1%2Cuf2",
+		},
+		{
+			name: "with max_results",
+			params: &types.ListInput{
+				ID:         "uid",
+				MaxResults: 10,
+			},
+			expect: endpointRoot + "uid" + "?max_results=10",
+		},
+		{
+			name: "with pagination_token",
+			params: &types.ListInput{
+				ID:              "uid",
+				PaginationToken: "ptoken",
+			},
+			expect: endpointRoot + "uid" + "?pagination_token=ptoken",
+		},
+		{
+			name: "all query parameters",
+			params: &types.ListInput{
+				Expansions:      fields.ExpansionList{"ex"},
+				ID:              "uid",
+				ListFields:      fields.ListFieldList{"lf"},
+				MaxResults:      10,
+				PaginationToken: "ptoken",
+				UserFields:      fields.UserFieldList{"uf"},
+			},
+			expect: endpointRoot + "uid" + "?expansions=ex&list.fields=lf&max_results=10&pagination_token=ptoken&user.fields=uf",
+		},
+		{
+			name: "has no required parameter",
+			params: &types.ListInput{
+				Expansions:      fields.ExpansionList{"ex"},
+				UserFields:      fields.UserFieldList{"uf"},
+				ListFields:      fields.ListFieldList{"lf"},
+				MaxResults:      10,
+				PaginationToken: "pagination_token",
+			},
+			expect: "",
+		},
+	}
+
+	for _, c := range cases {
+		t.Run(c.name, func(tt *testing.T) {
+			ep := c.params.ResolveEndpoint(endpointBase)
+			assert.Equal(tt, c.expect, ep)
+		})
+	}
+}
+
+func Test_ListMembersGet_Body(t *testing.T) {
+	cases := []struct {
+		name   string
+		params *types.ListInput
+	}{
+		{
+			name:   "empty params",
+			params: &types.ListInput{},
+		},
+		{
+			name:   "some params",
+			params: &types.ListInput{ID: "sid"},
+		},
+	}
+
+	for _, c := range cases {
+		t.Run(c.name, func(tt *testing.T) {
+			r, err := c.params.Body()
+			assert.NoError(tt, err)
+			assert.Nil(tt, r)
+		})
+	}
+}
+
+func Test_CreateInput_SetAccessToken(t *testing.T) {
+	cases := []struct {
+		name   string
+		token  string
+		expect string
+	}{
+		{
+			name:   "normal",
+			token:  "test-token",
+			expect: "test-token",
+		},
+		{
+			name:   "empty",
+			token:  "",
+			expect: "",
+		},
+	}
+
+	for _, c := range cases {
+		t.Run(c.name, func(tt *testing.T) {
+			p := &types.CreateInput{}
+			p.SetAccessToken(c.token)
+			assert.Equal(tt, c.expect, p.AccessToken())
+		})
+	}
+}
+
+func Test_CreateInput_ResolveEndpoint(t *testing.T) {
+	const endpointBase = "test/endpoint/"
+	cases := []struct {
+		name   string
+		params *types.CreateInput
+		expect string
+	}{
+		{
+			name:   "normal: only required parameter",
+			params: &types.CreateInput{ID: "lid"},
+			expect: endpointBase,
+		},
+		{
+			name:   "normal: has no required parameter",
+			params: &types.CreateInput{},
+			expect: "",
+		},
+	}
+
+	for _, c := range cases {
+		t.Run(c.name, func(tt *testing.T) {
+			ep := c.params.ResolveEndpoint(endpointBase)
+			assert.Equal(tt, c.expect, ep)
+		})
+	}
+}
+
+func Test_CreateInput_Body(t *testing.T) {
+	cases := []struct {
+		name   string
+		params *types.CreateInput
+		expect io.Reader
+	}{
+		{
+			name: "ok: has both of path and json parameters",
+			params: &types.CreateInput{
+				ID:     "lid",
+				UserID: "uid",
+			},
+			expect: strings.NewReader(`{"user_id":"uid"}`),
+		},
+		{
+			name: "ok: has no json parameters",
+			params: &types.CreateInput{
+				ID: "lid",
+			},
+			expect: strings.NewReader(`{"user_id":""}`),
+		},
+	}
+
+	for _, c := range cases {
+		t.Run(c.name, func(tt *testing.T) {
+			r, err := c.params.Body()
+			assert.NoError(tt, err)
+			assert.Equal(tt, c.expect, r)
+		})
+	}
+}
+
+func Test_CreateInput_ParameterMap(t *testing.T) {
+	cases := []struct {
+		name   string
+		params *types.CreateInput
+		expect map[string]string
+	}{
+		{
+			name:   "normal: only required parameter",
+			params: &types.CreateInput{ID: "id", UserID: "uid"},
+			expect: map[string]string{},
+		},
+		{
+			name:   "normal: has no required parameter",
+			params: &types.CreateInput{},
+			expect: map[string]string{},
+		},
+	}
+
+	for _, c := range cases {
+		t.Run(c.name, func(tt *testing.T) {
+			m := c.params.ParameterMap()
+			assert.Equal(tt, c.expect, m)
+		})
+	}
+}
+
+func Test_DeleteInput_SetAccessToken(t *testing.T) {
+	cases := []struct {
+		name   string
+		token  string
+		expect string
+	}{
+		{
+			name:   "normal",
+			token:  "test-token",
+			expect: "test-token",
+		},
+		{
+			name:   "empty",
+			token:  "",
+			expect: "",
+		},
+	}
+
+	for _, c := range cases {
+		t.Run(c.name, func(tt *testing.T) {
+			p := &types.DeleteInput{}
+			p.SetAccessToken(c.token)
+			assert.Equal(tt, c.expect, p.AccessToken())
+		})
+	}
+}
+
+func Test_DeleteInput_ResolveEndpoint(t *testing.T) {
+	const endpointRoot = "test/endpoint/"
+	const endpointBase = "test/endpoint/:id/:user_id"
+	cases := []struct {
+		name   string
+		params *types.DeleteInput
+		expect string
+	}{
+		{
+			name: "normal: only required parameter",
+			params: &types.DeleteInput{
+				ID:     "lid",
+				UserID: "uid",
+			},
+			expect: endpointRoot + "lid" + "/" + "uid",
+		},
+		{
+			name:   "normal: has no required parameter",
+			params: &types.DeleteInput{},
+			expect: "",
+		},
+	}
+
+	for _, c := range cases {
+		t.Run(c.name, func(tt *testing.T) {
+			ep := c.params.ResolveEndpoint(endpointBase)
+			assert.Equal(tt, c.expect, ep)
+		})
+	}
+}
+
+func Test_DeleteInput_Body(t *testing.T) {
+	cases := []struct {
+		name   string
+		params *types.DeleteInput
+		expect io.Reader
+	}{
+		{
+			name: "ok: has required parameters",
+			params: &types.DeleteInput{
+				ID:     "lid",
+				UserID: "uid",
+			},
+			expect: nil,
+		},
+		{
+			name:   "ok: has no required parameters",
+			params: &types.DeleteInput{},
+			expect: nil,
+		},
+	}
+
+	for _, c := range cases {
+		t.Run(c.name, func(tt *testing.T) {
+			r, err := c.params.Body()
+			assert.NoError(tt, err)
+			assert.Equal(tt, c.expect, r)
+		})
+	}
+}
+
+func Test_DeleteInput_ParameterMap(t *testing.T) {
+	cases := []struct {
+		name   string
+		params *types.DeleteInput
+		expect map[string]string
+	}{
+		{
+			name:   "normal: only required parameter",
+			params: &types.DeleteInput{ID: "id", UserID: "uid"},
+			expect: map[string]string{},
+		},
+		{
+			name:   "normal: has no required parameter",
+			params: &types.DeleteInput{},
+			expect: map[string]string{},
+		},
+	}
+
+	for _, c := range cases {
+		t.Run(c.name, func(tt *testing.T) {
+			m := c.params.ParameterMap()
+			assert.Equal(tt, c.expect, m)
+		})
+	}
+}
