+++ list/listtweetlookup/types/parameter.go
@@ -1,0 +1,88 @@
+package types
+
+import (
+	"io"
+	"net/url"
+	"strconv"
+	"strings"
+
+	"github.com/michimani/gotwi/fields"
+	"github.com/michimani/gotwi/internal/util"
+)
+
+type ListMaxResults int
+
+func (m ListMaxResults) Valid() bool {
+	return m > 1 && m <= 100
+}
+
+func (m ListMaxResults) String() string {
+	return strconv.Itoa(int(m))
+}
+
+type ListInput struct {
+	accessToken string
+
+	// Path parameter
+	ID string // List ID
+
+	// Query parameters
+	MaxResults      ListMaxResults
+	PaginationToken string
+	Expansions      fields.ExpansionList
+	TweetFields     fields.TweetFieldList
+	UserFields      fields.UserFieldList
+}
+
+var listQueryParameters = map[string]struct{}{
+	"max_results":      {},
+	"pagination_token": {},
+	"expansions":       {},
+	"tweet.fields":     {},
+	"user.fields":      {},
+}
+
+func (p *ListInput) SetAccessToken(token string) {
+	p.accessToken = token
+}
+
+func (p *ListInput) AccessToken() string {
+	return p.accessToken
+}
+
+func (p *ListInput) ResolveEndpoint(endpointBase string) string {
+	if p.ID == "" {
+		return ""
+	}
+
+	encoded := url.QueryEscape(p.ID)
+	endpoint := strings.Replace(endpointBase, ":id", encoded, 1)
+
+	pm := p.ParameterMap()
+	if len(pm) > 0 {
+		qs := util.QueryString(pm, listQueryParameters)
+		endpoint += "?" + qs
+	}
+
+	return endpoint
+}
+
+func (p *ListInput) Body() (io.Reader, error) {
+	return nil, nil
+}
+
+func (p *ListInput) ParameterMap() map[string]string {
+	m := map[string]string{}
+
+	if p.MaxResults.Valid() {
+		m["max_results"] = p.MaxResults.String()
+	}
+
+	if p.PaginationToken != "" {
+		m["pagination_token"] = p.PaginationToken
+	}
+
+	m = fields.SetFieldsParams(m, p.Expansions, p.TweetFields, p.UserFields)
+
+	return m
+}
