+++ list/pinnedlist/api.go
@@ -1,0 +1,47 @@
+package pinnedlist
+
+import (
+	"context"
+
+	"github.com/michimani/gotwi"
+	"github.com/michimani/gotwi/list/pinnedlist/types"
+)
+
+const (
+	listEndpoint   = "https://api.twitter.com/2/users/:id/pinned_lists"
+	createEndpoint = "https://api.twitter.com/2/users/:id/pinned_lists"
+	deleteEndpoint = "https://api.twitter.com/2/users/:id/pinned_lists/:list_id"
+)
+
+// Returns the Lists pinned by a specified user.
+// https://developer.twitter.com/en/docs/twitter-api/lists/pinned-lists/api-reference/get-users-id-pinned_lists
+func List(ctx context.Context, c *gotwi.Client, p *types.ListInput) (*types.ListOutput, error) {
+	res := &types.ListOutput{}
+	if err := c.CallAPI(ctx, listEndpoint, "GET", p, res); err != nil {
+		return nil, err
+	}
+
+	return res, nil
+}
+
+// Enables the authenticated user to pin a List.
+// https://developer.twitter.com/en/docs/twitter-api/lists/manage-lists/api-reference/post-users-id-pinned-lists
+func Create(ctx context.Context, c *gotwi.Client, p *types.CreateInput) (*types.CreateOutput, error) {
+	res := &types.CreateOutput{}
+	if err := c.CallAPI(ctx, createEndpoint, "POST", p, res); err != nil {
+		return nil, err
+	}
+
+	return res, nil
+}
+
+// Enables the authenticated user to unpin a List.
+// https://developer.twitter.com/en/docs/twitter-api/lists/manage-lists/api-reference/delete-users-id-pinned-lists-list_id
+func Delete(ctx context.Context, c *gotwi.Client, p *types.DeleteInput) (*types.DeleteOutput, error) {
+	res := &types.DeleteOutput{}
+	if err := c.CallAPI(ctx, deleteEndpoint, "DELETE", p, res); err != nil {
+		return nil, err
+	}
+
+	return res, nil
+}
