+++ list/pinnedlist/types/parameter.go
@@ -1,0 +1,143 @@
+package types
+
+import (
+	"encoding/json"
+	"io"
+	"net/url"
+	"strings"
+
+	"github.com/michimani/gotwi/fields"
+	"github.com/michimani/gotwi/internal/util"
+)
+
+type ListInput struct {
+	accessToken string
+
+	// Path parameter
+	ID string // User ID
+
+	// Query parameter
+	Expansions fields.ExpansionList
+	ListFields fields.ListFieldList
+	UserFields fields.UserFieldList
+}
+
+var listQueryParameters = map[string]struct{}{
+	"expansions":  {},
+	"list.fields": {},
+	"user.fields": {},
+}
+
+func (p *ListInput) SetAccessToken(token string) {
+	p.accessToken = token
+}
+
+func (p *ListInput) AccessToken() string {
+	return p.accessToken
+}
+
+func (p *ListInput) ResolveEndpoint(endpointBase string) string {
+	if p.ID == "" {
+		return ""
+	}
+
+	escaped := url.QueryEscape(p.ID)
+	endpoint := strings.Replace(endpointBase, ":id", escaped, 1)
+
+	pm := p.ParameterMap()
+	if len(pm) > 0 {
+		qs := util.QueryString(pm, listQueryParameters)
+		endpoint += "?" + qs
+	}
+
+	return endpoint
+}
+
+func (p *ListInput) Body() (io.Reader, error) {
+	return nil, nil
+}
+
+func (p *ListInput) ParameterMap() map[string]string {
+	m := map[string]string{}
+	m = fields.SetFieldsParams(m, p.Expansions, p.ListFields, p.UserFields)
+	return m
+}
+
+type CreateInput struct {
+	accessToken string
+
+	// Path parameter
+	ID string `json:"-"` // User ID
+
+	// JSON body parameter
+	ListID string `json:"list_id"` // required
+}
+
+func (p *CreateInput) SetAccessToken(token string) {
+	p.accessToken = token
+}
+
+func (p *CreateInput) AccessToken() string {
+	return p.accessToken
+}
+
+func (p *CreateInput) ResolveEndpoint(endpointBase string) string {
+	if p.ID == "" {
+		return ""
+	}
+
+	escaped := url.QueryEscape(p.ID)
+	endpoint := strings.Replace(endpointBase, ":id", escaped, 1)
+
+	return endpoint
+}
+
+func (p *CreateInput) Body() (io.Reader, error) {
+	json, err := json.Marshal(p)
+	if err != nil {
+		return nil, err
+	}
+
+	return strings.NewReader(string(json)), nil
+}
+
+func (p *CreateInput) ParameterMap() map[string]string {
+	return map[string]string{}
+}
+
+type DeleteInput struct {
+	accessToken string
+
+	// Path parameter
+	ID     string // User ID
+	ListID string
+}
+
+func (p *DeleteInput) SetAccessToken(token string) {
+	p.accessToken = token
+}
+
+func (p *DeleteInput) AccessToken() string {
+	return p.accessToken
+}
+
+func (p *DeleteInput) ResolveEndpoint(endpointBase string) string {
+	if p.ID == "" || p.ListID == "" {
+		return ""
+	}
+
+	escapedID := url.QueryEscape(p.ID)
+	endpoint := strings.Replace(endpointBase, ":id", escapedID, 1)
+	escapedListID := url.QueryEscape(p.ListID)
+	endpoint = strings.Replace(endpoint, ":list_id", escapedListID, 1)
+
+	return endpoint
+}
+
+func (p *DeleteInput) Body() (io.Reader, error) {
+	return nil, nil
+}
+
+func (p *DeleteInput) ParameterMap() map[string]string {
+	return map[string]string{}
+}
