+++ list/pinnedlist/types/response_test.go
@@ -1,0 +1,101 @@
+package types_test
+
+import (
+	"testing"
+
+	"github.com/michimani/gotwi/list/pinnedlist/types"
+	"github.com/michimani/gotwi/resources"
+	"github.com/stretchr/testify/assert"
+)
+
+func Test_ListOutput_HasPartialError(t *testing.T) {
+	cases := []struct {
+		name   string
+		res    *types.ListOutput
+		expect bool
+	}{
+		{
+			name:   "initial struct",
+			res:    &types.ListOutput{},
+			expect: false,
+		},
+		{
+			name: "has data",
+			res: &types.ListOutput{
+				Data: []resources.List{},
+			},
+			expect: false,
+		},
+	}
+
+	for _, c := range cases {
+		t.Run(c.name, func(tt *testing.T) {
+			hpe := c.res.HasPartialError()
+			assert.Equal(tt, c.expect, hpe)
+		})
+	}
+}
+
+func Test_CreateOutput_HasPartialError(t *testing.T) {
+	cases := []struct {
+		name   string
+		res    *types.CreateOutput
+		expect bool
+	}{
+		{
+			name:   "initial struct",
+			res:    &types.CreateOutput{},
+			expect: false,
+		},
+		{
+			name: "has data",
+			res: &types.CreateOutput{
+				Data: struct {
+					Pinned bool "json:\"pinned\""
+				}{
+					Pinned: false,
+				},
+			},
+			expect: false,
+		},
+	}
+
+	for _, c := range cases {
+		t.Run(c.name, func(tt *testing.T) {
+			hpe := c.res.HasPartialError()
+			assert.Equal(tt, c.expect, hpe)
+		})
+	}
+}
+
+func Test_DeleteOutput_HasPartialError(t *testing.T) {
+	cases := []struct {
+		name   string
+		res    *types.DeleteOutput
+		expect bool
+	}{
+		{
+			name:   "initial struct",
+			res:    &types.DeleteOutput{},
+			expect: false,
+		},
+		{
+			name: "has data",
+			res: &types.DeleteOutput{
+				Data: struct {
+					Pinned bool "json:\"pinned\""
+				}{
+					Pinned: false,
+				},
+			},
+			expect: false,
+		},
+	}
+
+	for _, c := range cases {
+		t.Run(c.name, func(tt *testing.T) {
+			hpe := c.res.HasPartialError()
+			assert.Equal(tt, c.expect, hpe)
+		})
+	}
+}
